refactor[#5]: conflict 제거
