+++ backend/src/main/java/com/handong/rebon/common/BaseEntity.java
@@ -1,14 +1,14 @@
 package com.handong.rebon.common;
 
-import lombok.Getter;
+import java.time.LocalDateTime;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
+import lombok.Getter;
 
 @Getter
 @MappedSuperclass
