--- backend/src/main/java/com/handong/rebon/shop/domain/tag/TagRepository.java
@@ -1,6 +1,0 @@
-package com.handong.rebon.shop.domain.tag;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TagRepository extends JpaRepository<Tag, Long> {
-}
