+++ backend/src/main/java/com/handong/rebon/shop/domain/tag/domain/repository/ShopTagRepository.java
@@ -1,7 +1,8 @@
 package com.handong.rebon.shop.domain.tag.domain.repository;
 
 import com.handong.rebon.shop.domain.tag.domain.ShopTag;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopTagRepository extends JpaRepository<ShopTag, Long>{
+public interface ShopTagRepository extends JpaRepository<ShopTag, Long> {
 }
