+++ backend/src/main/java/com/handong/rebon/shop/domain/tag/presentation/AdminTagController.java
@@ -1,9 +1,11 @@
 package com.handong.rebon.shop.domain.tag.presentation;
 
 import com.handong.rebon.shop.domain.tag.application.TagService;
-import lombok.RequiredArgsConstructor;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @RequestMapping("/admin/tags")
