+++ backend/src/test/java/com/handong/rebon/shop/domain/tag/domain/repository/TagRepositoryTest.java
@@ -1,7 +1,5 @@
 package com.handong.rebon.shop.domain.tag.domain.repository;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 class TagRepositoryTest {
 
 }
